+++ installer_prep.sh
@@ -14,7 +14,7 @@
 
 # Used to keep track of changes to 'linuxAIO.sh'.
 # Refer to the '[ Prepping ]' section of this script for more information.
-current_linuxAIO_revision="17"
+current_linuxAIO_revision="24"
 # Name of the master installer to be downloaded.
 # REASON: I placed the name in a variable just so there are no chances of accidentally
 #         misstyping the filename, in addition to it being slighlty shorter.
@@ -32,11 +32,16 @@ export _NC=$'\033[0m'
 export _GREY=$'\033[0;90m'
 export _CLRLN=$'\r\033[K'
 
-# PURPOSE: The '--no-hostname' flag for 'journalctl' only works with systemd 230 and
-#          later. So if systemd is older than 230, $_NO_HOSTNAME will not be created.
-if (($(journalctl --version | grep -oP "[0-9]+" | head -1) >= 230)) 2>/dev/null; then
-    export _NO_HOSTNAME="--no-hostname"
-fi
+## PURPOSE: The '--no-hostname' flag for 'journalctl' only works with systemd 230 and
+##          later. So if systemd is older than 230, $_NO_HOSTNAME will not be created.
+{
+    journalctl_version=$(journalctl --version) \
+        && journalctl_version=${journalctl_version:1:1}
+
+    if ((journalctl_version >= 230)) 2>/dev/null; then
+        export _NO_HOSTNAME="--no-hostname"
+    fi
+ } 2>/dev/null
 
 
 #### End of [ Exported and/or Globally Used Variables ]
@@ -84,7 +89,7 @@ linuxAIO_update() {
     ####
     # Function Info
     # -------------
-    # Download the latest version of 'linuxAIO.sh' if $_LINUXAIO_REVISION and 
+    # Download the latest version of 'linuxAIO.sh' if $_LINUXAIO_REVISION and
     # $current_linuxAIO_revision aren't of equal value.
     #
     # Purpose
@@ -104,43 +109,80 @@ linuxAIO_update() {
     # 'linuxAIO.sh'.
     ####
 
-    ## Save the value of $installer_branch, $allow_run_as_root, $_NADEKO_INSTALL_VERSION
-    ## specified in 'linuxAIO.sh', to be set in the new 'linuxAIO.sh'.
+    ## Save the values of the Configuration Variables specified in 'linuxAIO.sh', to be
+    ## set in the new 'linuxAIO.sh'.
     ## NOTE: Declaration and intialation is seperated at the recommendation by
     ##       shellcheck.
-    local installer_branch
-    local allow_run_as_root
-    local nadeko_install_version
+    local installer_branch              # $installer_branch
+    local installer_branch_found        # $installer_branch
     installer_branch=$(grep '^installer_branch=.*' linuxAIO.sh)
+    installer_branch_found=$?
+    local allow_run_as_root             # $allow_run_as_root
+    local allow_run_as_root_found       # $allow_run_as_root
     allow_run_as_root=$(grep '^allow_run_as_root=.*' linuxAIO.sh)
+    allow_run_as_root_found=$?
+    local nadeko_install_version        # $_NADEKO_INSTALL_VERSION
+    local nadeko_install_version_found  # $_NADEKO_INSTALL_VERSION
     nadeko_install_version=$(grep '^export _NADEKO_INSTALL_VERSION=.*' linuxAIO.sh)
+    nadeko_install_version_found=$?
 
     echo "$_YELLOW'linuxAIO.sh' is not up to date$_NC"
     echo "Downloading latest 'linuxAIO.sh'..."
     curl -O "$_RAW_URL"/linuxAIO.sh
+    sudo chmod +x linuxAIO.sh
 
-    echo "Applying set configurations to 'linuxAIO.sh'..."
+    echo "Applying existing configurations to the new 'linuxAIO.sh'..."
+
+    ####################################################################################
+    #### [[ $installer_branch ]]
+
+
     ## Set $installer_branch inside of the new 'linuxAIO.sh'.
-    # Sed for linux.
-    sed -i "s/^installer_branch=.*/$installer_branch/" linuxAIO.sh ||
-        # Sed for macOS.
-        sed -i '' "s/^installer_branch=.*/$installer_branch/" linuxAIO.sh
-    
+    if [[ $installer_branch_found = 0 ]]; then
+        # Sed for linux || Sed for macOS.
+        sed -i "s/^installer_branch=.*/$installer_branch/" linuxAIO.sh \
+            || sed -i '' "s/^installer_branch=.*/$installer_branch/" linuxAIO.sh
+    fi
+
+
+    #### End of [[ $installer_branch ]]
+    ####################################################################################
+    #### [[ $allow_run_as_root ]]
+
+
     ## Set $allow_run_as_root inside of the new 'linuxAIO.sh'.
-    # Sed for linux.
-    sed -i "s/^allow_run_as_root=.*/$allow_run_as_root/" linuxAIO.sh ||
-        # Sed for macOS.
-        sed -i '' "s/^allow_run_as_root=.*/$allow_run_as_root/" linuxAIO.sh
+    if [[ $allow_run_as_root_found = 0 ]]; then
+        # Sed for linux || Sed for macOS.
+        sed -i "s/^allow_run_as_root=.*/$allow_run_as_root/" linuxAIO.sh \
+            || sed -i '' "s/^allow_run_as_root=.*/$allow_run_as_root/" linuxAIO.sh
+    fi
+
+
+    #### End of [[ $allow_run_as_root ]]
+    ####################################################################################
+    #### [[ $_NADEKO_INSTALL_VERSION ]]
+
 
     ## Set $nadeko_install_version inside of the new 'linuxAIO.sh'.
-    # Sed for linux.
-    sed -i "s/^export _NADEKO_INSTALL_VERSION=.*/$nadeko_install_version/" linuxAIO.sh ||
-        # Sed for macOS.
-        sed -i '' "s/^export _NADEKO_INSTALL_VERSION=.*/$nadeko_install_version/" linuxAIO.sh
-    
-    sudo chmod +x linuxAIO.sh
-    echo "${_GREEN}Successfully downloaded the newest version of 'linuxAIO.sh'$_NC"
-    _CLEAN_EXIT "0" "Exiting" "true"
+    if [[ $nadeko_install_version_found = 0 ]]; then
+        # Sed for linux || Sed for macOS.
+        sed -i "s/^export _NADEKO_INSTALL_VERSION=.*/$nadeko_install_version/" linuxAIO.sh \
+            || sed -i '' "s/^export _NADEKO_INSTALL_VERSION=.*/$nadeko_install_version/" linuxAIO.sh
+    fi
+
+
+    #### End of [[ $_NADEKO_INSTALL_VERSION ]]
+    ####################################################################################
+    #### [[ Finishing up ]]
+
+
+    echo "${_GREEN}Successfully downloaded the newest version of 'linuxAIO.sh' and" \
+        "applied changes to the newest version of 'linuxAIO.sh'$_NC"
+    clean_up "0" "Exiting" "true"
+
+
+    #### End of [[ Finishing up ]]
+    ####################################################################################
 }
 
 unsupoorted() {
@@ -153,17 +195,17 @@ unsupoorted() {
 
     echo "${_RED}Your operating system/Linux Distribution is not OFFICIALLY supported" \
         "for the installation, setup, and/or use of NadekoBot" >&2
-    echo "${_YELLOW}WARNING: By continuing, you accept that unexpected behaviors" \
+    echo "${_YELLOW}WARNING: By continuing, you accept that unexpected behaviors"     \
         "may occur. If you run into any errors or problems with the installation and" \
-        "use of the NadekoBot, you are on your own. We do not provide support for" \
+        "use of the NadekoBot, you are on your own. We do not provide support for"    \
         "distributions that we don't officially support.$_NC"
     read -rp "Would you like to continue anyways? [y/N] " choice
     # Convert user input to lowercase.
     choice=$(echo "$choice" | tr '[:upper:]' '[:lower:]')
     case "$choice" in
         y|yes) clear -x; execute_master_installer ;;
-        n|no)  _CLEAN_EXIT "0" "Exiting" ;;
-        *)     _CLEAN_EXIT "0" "Exiting" ;;
+        n|no)  clean_up "0" "Exiting" ;;
+        *)     clean_up "0" "Exiting" ;;
     esac
 }
 
@@ -175,15 +217,12 @@ execute_master_installer() {
     ####
 
     _DOWNLOAD_SCRIPT "$master_installer" "$master_installer" "true"
+    # TODO: Perhaps remove one of the duplicate code snippets.
     ./nadeko_master_installer.sh
+    clean_up "$?" "Exiting"
 }
 
-
-########################################################################################
-#### [[ Functions To Be Exported ]]
-
-
-_CLEAN_EXIT() {
+clean_up() {
     ####
     # Function Info
     # -------------
@@ -201,22 +240,35 @@ _CLEAN_EXIT() {
         "prereqs_installer.sh" "nadeko_latest_installer.sh" "nadeko_runner.sh"
         "nadeko_master_installer.sh")
 
-    ### PURPOSE: Sometimes the output requires the use of a new-line symbol to seperate 
+    ### PURPOSE: Sometimes the output requires the use of a new-line symbol to seperate
     ###          the previous text, though sometimes it doesn't.
-    if [[ $3 = true ]]; then
-        echo "Cleaning up..."
-    else
-        echo -e "\nCleaning up..."
+    if [[ $3 = true ]]; then echo "Cleaning up..."
+    else                     echo -e "\nCleaning up..."
     fi
+
+    cd "$_WORKING_DIR" || {
+        echo "${_RED}Failed to move to project root directory$_NC" >&2
+        exit 1
+	}
+
+	# TODO: Add echo stuff for output.
+    if [[ -d NadekoTMPDir ]]; then rm -rf NadekoTMPDir
+	fi
 
     ## Remove any and all files specified in $installer_files.
     for file in "${installer_files[@]}"; do
-        if [[ -f $file ]]; then rm "$file"; fi
+        if [[ -f $file ]]; then rm "$file"
+        fi
     done
 
     echo "$2..."
     exit "$1"
 }
+
+
+########################################################################################
+#### [[ Functions To Be Exported ]]
+
 
 _DOWNLOAD_SCRIPT() {
     ####
@@ -225,11 +277,11 @@ _DOWNLOAD_SCRIPT() {
     # Download the specified script and modify it's execution permissions.
     #
     # @param $1 Name of script to donwload.
-    # @param $2 Name to rename $1 with. 
+    # @param $2 Name to rename $1 with.
     ####
 
-    if [[ ! $3 ]]; then echo "Downloading '$1'..."; fi
-
+    if [[ ! $3 ]]; then echo "Downloading '$1'..."
+    fi
     curl -s "$_RAW_URL"/"$1" -o "$2"
     sudo chmod +x "$2"
 }
@@ -246,7 +298,7 @@ _DOWNLOAD_SCRIPT() {
 # Execute when the user uses 'Ctrl + Z', 'Ctrl + C', or otherwise forcefully exits the
 # installer.
 trap 'echo -e "\n\nScript forcefully stopped"
-    _CLEAN_EXIT "1" "Exiting" "true"' \
+    clean_up "2" "Exiting" "true"' \
     SIGINT SIGTSTP SIGTERM
 
 
@@ -255,26 +307,26 @@ trap 'echo -e "\n\nScript forcefully stopped"
 #### [ Prepping ]
 
 
-## If the current 'linuxAIO.sh' revision is number 9 or later...
+# If the current 'linuxAIO.sh' revision is number 9 or later...
 if [[ $_LINUXAIO_REVISION && $_LINUXAIO_REVISION != "$current_linuxAIO_revision" ]]; then
     linuxAIO_update
-## If the current 'linuxAIO.sh' revision is number 8 or earlier...
+# If the current 'linuxAIO.sh' revision is number 8 or earlier...
 elif [[ $linuxAIO_revision && $linuxAIO_revision != "$current_linuxAIO_revision" ]]; then
     echo "$_YELLOW'linuxAIO.sh' is not up to date"
     echo "${_CYAN}Due to some breaking changes between revision 8 and 9 you are" \
-        "required to manually download the newest version of" \
-        "'linuxAIO.sh'. You can do so by executing the following:"
+        "required to manually download the newest version of 'linuxAIO.sh'. You can" \
+        "do so by executing the following:"
     echo "    mv linuxAIO.sh linuxAIO.sh.old && curl -O" \
         "https://raw.githubusercontent.com/$installer_repo/$installer_branch/linuxAIO.sh" \
         "&& sudo chmod +x linuxAIO.sh$_NC"
-    _CLEAN_EXIT "0" "Exiting" "true"
+    clean_up "0" "Exiting" "true"
 fi
 
 # Change the working directory to the location of the executed scrpt.
 cd "$(dirname "$0")" || {
     echo "${_RED}Failed to change working directory" >&2
     echo "${_CYAN}Change your working directory to that of the executed script$_NC"
-    _CLEAN_EXIT "1" "Exiting" "true"
+    clean_up "1" "Exiting" "true"
 }
 
 export _WORKING_DIR="$PWD"
@@ -290,57 +342,56 @@ clear -x  # Clear the screen of any text.
 
 detect_sys_info
 export _DISTRO _SVER _VER _ARCH
-export -f _CLEAN_EXIT _DOWNLOAD_SCRIPT
+export -f _DOWNLOAD_SCRIPT
 
-echo "SYSTEM INFO"
-echo "Bit Type: $bits"
-echo "Architecture: $_ARCH"
-printf "Distro: "
-# Use $_DISTRO if $pname is empty.
-if [[ -n $pname ]]; then echo "$pname"; else echo "$_DISTRO"; fi
-echo "Distro Version: $_VER"
-echo ""
+# Use $_DISTRO if $pname is unset or null.
+echo "SYSTEM INFO
+Bit Type: $bits
+Architecture: $_ARCH
+Distro: ${pname:=$_DISTRO}
+Distro Version: $_VER
+"
 
 ### Check if the operating system is supported by NadekoBot and installer.
-if [[ $_DISTRO = "ubuntu" ]]; then
-    if [[ $bits = 64 ]]; then  # A.1.
+if [[ $bits = 64 ]]; then
+    # Ubuntu:
+    #   16.04
+    #   18.04
+    #   20.04
+    if [[ $_DISTRO = "ubuntu" ]]; then
         case "$_VER" in
-            16.04) execute_master_installer ;;
-            18.04) execute_master_installer ;;
-            20.04) execute_master_installer ;;
-            *)     unsupoorted ;;
+            16.04|18.04|20.04) execute_master_installer ;;
+            *)                 unsupoorted ;;
         esac
-    else
-        unsupoorted
-    fi
-elif [[ $_DISTRO = "debian" ]]; then
-    if [[ $bits = 64 ]]; then  # A.1.
+    # Debian:
+    #   9
+    #   10
+    elif [[ $_DISTRO = "debian" ]]; then
         case "$_SVER" in
-            9)  execute_master_installer ;;
-            10) execute_master_installer ;;
-            *)  unsupoorted ;;
+            9|10) execute_master_installer ;;
+            *)    unsupoorted ;;
         esac
-    else
-        unsupoorted
-    fi
-elif [[ $_DISTRO = "linuxmint" ]]; then
-    if [[ $bits = 64 ]]; then  # A.1.
+    # Linux Mint:
+    #   18
+    #   19
+    #   20
+    elif [[ $_DISTRO = "linuxmint" ]]; then
         case "$_SVER" in
-            18) execute_master_installer ;;
-            19) execute_master_installer ;;
-            20) execute_master_installer ;;
-            *)  unsupoorted ;;
+            18|19|20) execute_master_installer ;;
+            *)        unsupoorted ;;
+        esac
+    # macOS:
+    #   10.14
+    #   10.15
+    #   11.*
+    elif [[ $_DISTRO = "Darwin" ]]; then
+        case "$_SVER" in
+            10.14|10.15|11.*) execute_master_installer ;;
+            *)                unsupoorted ;;
         esac
     else
         unsupoorted
     fi
-elif [[ $_DISTRO = "Darwin" ]]; then
-    case "$_SVER" in
-        10.14) execute_master_installer ;;
-        10.15) execute_master_installer ;;
-        11.*)  execute_master_installer ;;
-        *)     unsupoorted ;;
-    esac
 else
     unsupoorted
 fi