+++ linux_nadeko_runner.sh
@@ -10,7 +10,7 @@
 timer=60
 # Save the current time and date, which will be used in conjunction with 'journalctl'.
 start_time=$(date +"%F %H:%M:%S")
-nadeko_service_content="[Unit] 
+nadeko_service_content="[Unit]
 Description=NadekoBot service
 
 [Service]
@@ -40,34 +40,30 @@ fi
 
 
 # Check if the service exists.
-if [[ -f $_NADEKO_SERVICE ]]; then
-    echo "Updating '$_NADEKO_SERVICE_NAME'..."
-else
-    echo "Creating '$_NADEKO_SERVICE_NAME'..."
+if [[ -f $_NADEKO_SERVICE ]]; then echo "Updating '$_NADEKO_SERVICE_NAME'..."
+else                               echo "Creating '$_NADEKO_SERVICE_NAME'..."
 fi
 
 # Create/update '$_NADEKO_SERVICE_NAME'.
-echo "$nadeko_service_content" | sudo tee "$_NADEKO_SERVICE" &>/dev/null &&
-        sudo systemctl daemon-reload || {
+echo "$nadeko_service_content" | sudo tee "$_NADEKO_SERVICE" &>/dev/null \
+        && sudo systemctl daemon-reload || {
     echo "${_RED}Failed to create '$_NADEKO_SERVICE_NAME'" >&2
     echo "${_CYAN}This service must exist for NadekoBot to work$_NC"
     read -rp "Press [Enter] to return to the installer menu"
-    exit 1
+    exit 4
 }
 
 ## Disable or enable '$_NADEKO_SERVICE_NAME'.
 echo "$dis_en_upper '$_NADEKO_SERVICE_NAME'..."
 sudo systemctl "$dis_en_lower" "$_NADEKO_SERVICE_NAME" || {
     echo "${_RED}Failed to $dis_en_lower '$_NADEKO_SERVICE_NAME'" >&2
-    echo "${_CYAN}This service must be ${dis_en_lower}d in order to use this" \
-        "run mode$_NC"
+    echo "${_CYAN}This service must be ${dis_en_lower}d in order to use this run mode$_NC"
     read -rp "Press [Enter] to return to the installer menu"
-    exit 1
+    exit 4
 }
 
 # Check if 'NadekoRun.sh' exists.
-if [[ -f NadekoRun.sh ]]; then
-    echo "Updating 'NadekoRun.sh'..."
+if [[ -f NadekoRun.sh ]]; then echo "Updating 'NadekoRun.sh'..."
 ## Create 'NadekoRun.sh' if it doesn't exist.
 else
     echo "Creating 'NadekoRun.sh'..."
@@ -77,45 +73,48 @@ fi
 
 ## Add the code required to run NadekoBot in the background, to 'NadekoRun.sh'.
 if [[ $_CODENAME = "NadekoRun" ]]; then
-    echo -e "#!bin/bash
-        \n \
-        \n_code_name_=\"NadekoRun\" \
-        \n \
-        \necho \"Running NadekoBot in the background\" \
-        \nyoutube-dl -U \
-        \n \
-        \ncd $_WORKING_DIR/NadekoBot \
-        \ndotnet build -c Release \
-        \ncd $_WORKING_DIR/NadekoBot/src/NadekoBot \
-        \necho \"Running NadekoBot...\" \
-        \ndotnet run -c Release \
-        \necho \"Done\" \
-        \ncd $_WORKING_DIR \
-        \n" > NadekoRun.sh
+    printf '%s\n' \
+        "#!bin/bash" \
+        "" \
+        "_code_name_=\"NadekoRun\"" \
+        "" \
+        "echo \"Running NadekoBot in the background\"" \
+        "youtube-dl -U" \
+        "" \
+        "cd $_WORKING_DIR/NadekoBot" \
+        "dotnet build -c Release" \
+        "cd $_WORKING_DIR/NadekoBot/src/NadekoBot" \
+        "echo \"Running NadekoBot...\"" \
+        "dotnet run -c Release" \
+        "echo \"Done\"" \
+        "cd $_WORKING_DIR" \
+        "" > NadekoRun.sh
 ## Add code required to run NadekoBot in the background with auto restart, to
 ## 'NadekoRun.sh'.
 else
-    echo -e "#!/bin/bash \
-        \n \
-        \n_code_name_=\"NadekoRunAR\" \
-        \n \
-        \necho \"\" \
-        \necho \"Running NadekoBot in the background with auto restart\" \
-        \nyoutube-dl -U \
-        \n \
-        \nsleep 5 \
-        \ncd $_WORKING_DIR/NadekoBot \
-        \ndotnet build -c Release \
-        \n \
-        \nwhile true; do \
-        \n    cd $_WORKING_DIR/NadekoBot/src/NadekoBot && \
-        \n        dotnet run -c Release \
-        \n \
-        \n    youtube-dl -U \
-        \n    sleep 10 \
-        \ndone \
-        \n \
-        \necho \"Stopping NadekoBot\"" > NadekoRun.sh
+    printf '%s\n' \
+        "#!/bin/bash" \
+        "" \
+        "_code_name_=\"NadekoRunAR\"" \
+        "" \
+        "echo \"\"" \
+        "echo \"Running NadekoBot in the background with auto restart\"" \
+        "youtube-dl -U" \
+        "" \
+        "sleep 5" \
+        "cd $_WORKING_DIR/NadekoBot" \
+        "dotnet build -c Release" \
+        "" \
+        "while true; do" \
+        "    cd $_WORKING_DIR/NadekoBot/src/NadekoBot &&" \
+        "        dotnet run -c Release" \
+        "" \
+        "    youtube-dl -U" \
+        "    sleep 10" \
+        "done" \
+        "" \
+        "echo \"Stopping NadekoBot\"" \
+        "" > NadekoRun.sh
 fi
 
 ## Restart $_NADEKO_SERVICE_NAME if it is currently running.
@@ -124,7 +123,7 @@ if [[ $_NADEKO_SERVICE_STATUS = "active" ]]; then
     sudo systemctl restart "$_NADEKO_SERVICE_NAME" || {
         echo "${_RED}Failed to restart '$_NADEKO_SERVICE_NAME'$_NC" >&2
         read -rp "Press [Enter] to return to the installer menu"
-        exit 1
+        exit 4
     }
     echo "Waiting $timer seconds for '$_NADEKO_SERVICE_NAME' to restart..."
 ## Start $_NADEKO_SERVICE_NAME if it is NOT currently running.
@@ -133,7 +132,7 @@ else
     sudo systemctl start "$_NADEKO_SERVICE_NAME" || {
         echo "${_RED}Failed to start '$_NADEKO_SERVICE_NAME'$_NC" >&2
         read -rp "Press [Enter] to return to the installer menu"
-        exit 1
+        exit 4
     }
     echo "Waiting $timer seconds for '$_NADEKO_SERVICE_NAME' to start..."
 fi
@@ -145,10 +144,10 @@ while ((timer > 0)); do
     ((timer-=1))
 done
 
-# NOTE: $_NO_HOSTNAME is purposefully unquoted. Do not quote the variable.
-echo -e "\n\n-------- $_NADEKO_SERVICE_NAME startup logs ---------" \
-    "\n$(journalctl -q -u nadeko -b $_NO_HOSTNAME -S "$start_time" 2>/dev/null ||
-    sudo journalctl -q -u nadeko -b $_NO_HOSTNAME -S "$start_time")" \
+# NOTE: $_NO_HOSTNAME is purposefully unquoted. Do not quote it!
+echo -e "\n\n-------- $_NADEKO_SERVICE_NAME startup logs ---------"            \
+    "\n$(journalctl -q -u nadeko -b $_NO_HOSTNAME -S "$start_time" 2>/dev/null \
+         || sudo journalctl -q -u nadeko -b $_NO_HOSTNAME -S "$start_time")"   \
     "\n--------- End of $_NADEKO_SERVICE_NAME startup logs --------\n"
 
 echo -e "${_CYAN}Please check the logs above to make sure that there aren't any" \