+++ nadeko_latest_installer.sh
@@ -24,8 +24,7 @@ netcoreapp_version="netcoreapp2.1"
 
 ## NOTE: 'cp' on macOS doesn't have the flag options "RT", while the Linux version does.
 ## Use the "RT" flags if the OS is NOT macOS.
-if [[ $_DISTRO != "Darwin" ]]; then
-    cp_flag="RT"
+if [[ $_DISTRO != "Darwin" ]]; then cp_flag="RT"
 ## If running on macOS, use the "r" flag. But if 'gcp' (GNU cp) is installed, then use
 ## the "RT" flags and make 'cp' an alias of 'gcp'
 else
@@ -42,74 +41,6 @@ fi
 
 
 #### End of [ Variables ]
-########################################################################################
-#### [ Functions ]
-
-
-clean_up() {
-    ####
-    # Function Info
-    # -------------
-    # If 'nadeko_latest_installer.sh', for whatever reason, is not able to completely
-    # update or download NadekoBot, this function will cleanup any remaining files and
-    # restore the previous code and configurations.
-    #
-    # @param $1 Determines whether or not this function kills the script's parent
-    #           processes.
-    ####
-
-    # Files to be removed.
-    local installer_files=("credentials_setup.sh" "installer_prep.sh"
-        "prereqs_installer.sh" "nadeko_latest_installer.sh"
-        "nadeko_master_installer.sh" "nadeko_runner.sh")
-
-    echo "Cleaning up files and directories..."
-    ## Remove any and all files specified in $installer_files.
-    for file in "${installer_files[@]}"; do
-        if [[ -f $_WORKING_DIR/$file ]]; then rm "$_WORKING_DIR"/"$file"; fi
-    done
-
-    ## Remove the NadekoBot code that was just downloaded, then restore the old code, if
-    ## both of the 'NadekoBot' and 'NadekoBot.bak' directories exist.
-    if [[ (-d $_WORKING_DIR/NadekoBot || -d $_WORKING_DIR/NadekoTMPDir) && 
-            -d $_WORKING_DIR/NadekoBot.bak ]]; then
-        rm -rf "$_WORKING_DIR"/NadekoBot 2>/dev/null
-        rm -rf "$_WORKING_DIR"/NadekoTMPDir 2>/dev/null
-
-        echo "Restoring from 'NadekoBot.bak'..."
-        mv -f "$_WORKING_DIR"/NadekoBot.bak "$_WORKING_DIR"/NadekoBot || {
-            echo "${_RED}Failed to restore from 'NadekoBot.bak'" >&2
-            echo "${_CYAN}Manually rename 'NadekoBot.bak' to 'NadekoBot'$_NC"
-        }
-    fi
-
-    ## Kill the script's parent processes, if param $1 is true.
-    if [[ $1 = true ]]; then
-        echo "Killing parent processes..."
-        kill -9 "$_NADEKO_MASTER_INSTALLER_PID" "$_INSTALLER_PREP_PID"
-        echo "Exiting..."
-        exit 1
-    fi
-}
-
-
-#### End of [ Functions ]
-########################################################################################
-#### [ Error Trapping ]
-
-
-# Execute when the user uses 'Ctrl + Z', 'Ctrl + C', or otherwise forcefully exits the
-# installer.
-trap 'echo -e "\n\nScript forcefully stopped"
-    clean_up
-    echo "Killing parent processes..."
-    kill -9 "$_NADEKO_MASTER_INSTALLER_PID" "$_INSTALLER_PREP_PID"
-    echo "Exiting..."
-    exit 1' \
-    SIGINT SIGTSTP SIGTERM
-
-
-#### End of [ Error Trapping ]
 ########################################################################################
 #### [ Main ]
 
@@ -132,12 +63,13 @@ fi
 ########################################################################################
 #### [[ Create Backup, Then Update ]]
 
+
 ## A.1.
 ## Create a temporary folder to download NadekoBot into.
 mkdir NadekoTMPDir
 cd NadekoTMPDir || {
     echo "${_RED}Failed to change working directory$_NC" >&2
-    clean_up "true"
+    exit 1
 }
 
 echo "Downloading NadekoBot..."
@@ -145,14 +77,14 @@ echo "Downloading NadekoBot..."
 git clone -b "$_NADEKO_INSTALL_VERSION" --recursive --depth 1 \
         https://gitlab.com/Kwoth/NadekoBot || {
     echo "${_RED}Failed to download NadekoBot$_NC" >&2
-    clean_up "true"
+    exit 1
 }
 
 # If $_DISTRO isn't Darwin and '/tmp/NuGetScratch' exists...
 if [[ -d /tmp/NuGetScratch && $_DISTRO != "Darwin" ]]; then
     echo "Modifying ownership of '/tmp/NuGetScratch' and '/home/$USER/.nuget'"
     # Due to permission errors cropping up every now and then, especially when the
-    # installer is executed with root privilege, it's neccessary to make sure that 
+    # installer is executed with root privilege, it's neccessary to make sure that
     # '/tmp/NuGetScratch' and '/home/$USER/.nuget' are owned by the user that the
     # installer is currently being run under.
     sudo chown -R "$USER":"$USER" /tmp/NuGetScratch /home/"$USER"/.nuget || {
@@ -168,20 +100,20 @@ fi
 echo "Building NadekoBot..."
 cd NadekoBot || {
     echo "${_RED}Failed to change working directory$_NC" >&2
-    clean_up "true"
+    exit 1
 }
 
 dotnet build --configuration Release || {
     echo "${_RED}Failed to build NadekoBot$_NC" >&2
-    clean_up "true"
+    exit 1
 }
 cd "$_WORKING_DIR" || {
     echo "${_RED}Failed to return to the project's root directory$_NC" >&2
-    clean_up "true"
+    exit 1
 }
 
 ## Move credentials, database, and other data to the new version of NadekoBot.
-if [[ -d NadekoBot.old && -d NadekoBot || ! -d NadekoBot.old && -d NadekoBot ]]; then
+if [[ -d NadekoBot ]]; then
     echo "Copping 'credentials.json' to new version..."
     cp -f "$current_credentials" "$new_credentials" &>/dev/null
     echo "Copping database to the new version..."
@@ -190,22 +122,22 @@ if [[ -d NadekoBot.old && -d NadekoBot || ! -d NadekoBot.old && -d NadekoBot ]];
     ## Check if an old netcoreapp version exists, then move the database within it, to
     ## the new netcorapp version.
     while read -r netcoreapp; do
-        if [[ $netcoreapp != "$netcoreapp_version" &&
-                -f "$new_database"/Release/"$netcoreapp"/data/NadekoBot.db ]]; then
+        if [[ $netcoreapp != "$netcoreapp_version" \
+                && -f "$new_database"/Release/"$netcoreapp"/data/NadekoBot.db ]]; then
             echo "${_YELLOW}WARNING: Old netcoreapp version detected$_NC"
             echo "Moving database to new netcoreapp version..."
 
             cp "$new_database"/Release/"$netcoreapp"/data/NadekoBot.db \
                     "$new_database"/Release/"$netcoreapp_version"/data/NadekoBot.db || {
                 echo "${_RED}Failed to move database$_NC" >&2
-                clean_up "true"
+                exit 1
             }
 
             echo "Removing '$netcoreapp'..."
             rm -rf "$new_database"/Release/"$netcoreapp" || {
                 echo "${_RED}Failed to remove '$netcoreapp'" >&2
-                echo -e "${_CYAN}Please manually remove '$netcoreapp' before continuing" \
-                    "\nLocation: $_WORKING_DIR/$new_database/Release/$netcoreapp$_NC"
+                echo "${_CYAN}Please manually remove '$netcoreapp' before continuing"
+                echo "Location: $_WORKING_DIR/$new_database/Release/$netcoreapp$_NC"
             }
         fi
     done < <(ls "$_WORKING_DIR"/"$new_database"/Release/)